# WiFiServerSecure: Cache SSL sessions