--- a/libraries/ESP8266WiFi/src/BearSSLHelpers.h
+++ b/libraries/ESP8266WiFi/src/BearSSLHelpers.h
@@ -144,6 +144,43 @@ class Session {
     br_ssl_session_parameters _session;
 };
 
+// Cache for the TLS sessions of multiple clients.
+// Use with BearSSL::WiFiServerSecure::setCache
+class ServerSessions {
+  friend class WiFiClientSecureCtx;
+
+  public:
+    // Uses the given buffer of the given size to store the size and initializes it.
+    ServerSessions(uint8_t *store, uint32_t size) : ServerSessions(store, size, false) {}
+
+    // Dynamically allocates a buffer of the given size to store the cache and initializes it.
+    // If the allocation of the buffer wasn't successfull, the value
+    // returned by size() will be 0.
+    ServerSessions(uint32_t size) : ServerSessions(size > 0 ? new uint8_t[size] : nullptr, size, true) {}
+
+    ~ServerSessions();
+
+    // Returns the size of the store for the cache.
+    uint32_t size() { return _size; }
+
+  private:
+    ServerSessions(uint8_t *store, uint32_t size, bool isDynamic);
+
+    // Returns the cache's vtable or null if the cache has no capacity.
+    const br_ssl_session_cache_class **getCache();
+
+    // Size of the store.
+    uint32_t _size;
+    // Store where the informations for the sessions are stored.
+    uint8_t *_store;
+    // Whether the store is dynamically allocated.
+    // If this is true, the store needs to be freed in the destructor.
+    bool _isDynamic;
+
+    // Cache of the server using the _store.
+    br_ssl_session_cache_lru _cache;
+};
+
 // Updater SHA256 hash and signature verification
 class HashSHA256 : public UpdaterHashClass {
   public:
@@ -170,7 +207,7 @@ class SigningVerifier : public UpdaterVerifyClass {
   private:
     PublicKey *_pubKey;
 };
-  
+
 // Stack thunked versions of calls
 extern "C" {
 extern unsigned char *thunk_br_ssl_engine_recvapp_buf( const br_ssl_engine_context *cc, size_t *len);